requirements.txt
streamlit==1.29.0
google-generativeai==0.3.2
sentence-transformers==2.2.2
numpy==1.24.0
pandas==2.0.0
plotly==5.13.0
python-dotenv==1.0.0
tiktoken>=0.5.0
pytest>=7.3.1
black>=23.3.0
flake8>=6.0.0
isort>=5.12.0
pre-commit>=3.3.3
coverage>=7.2.7
pytest-cov>=4.1.0
mypy>=1.3.0
python-multipart>=0.0.6
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
redis>=4.6.0 